# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hamza Cafe API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MenuItem {
  id: number;
  name: string;
  nameAr: string;
  price: number;
  category: string;
  /** @nullable */
  description?: string | null;
  available?: boolean;
}

export interface Category {
  name: string;
  nameAr: string;
  count: number;
}

export interface OrderItem {
  menuItemId: number;
  quantity: number;
  /** @nullable */
  notes?: string | null;
}

export interface OrderInput {
  customerName: string;
  phone: string;
  items: OrderItem[];
  /** @nullable */
  notes?: string | null;
}

export interface OrderItemDetail {
  menuItemId: number;
  menuItemName: string;
  menuItemNameAr: string;
  quantity: number;
  unitPrice: number;
  subtotal: number;
  /** @nullable */
  notes?: string | null;
}

export interface Order {
  id: number;
  customerName: string;
  phone: string;
  status: string;
  total: number;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  items: OrderItemDetail[];
}

export type OrderStatsPopularItemsItem = {
  name: string;
  nameAr: string;
  count: number;
};

export interface OrderStats {
  totalOrders: number;
  totalRevenue: number;
  popularItems: OrderStatsPopularItemsItem[];
}

export type ListMenuItemsParams = {
category?: string;
};