# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hamza Cafe API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: menu
    description: Menu items operations
  - name: orders
    description: Order operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /menu:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List all menu items
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"

  /menu/categories:
    get:
      operationId: listCategories
      tags: [menu]
      summary: List all categories with item counts
      responses:
        "200":
          description: Categories with counts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /menu/{id}:
    get:
      operationId: getMenuItem
      tags: [menu]
      summary: Get a single menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found

  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Bad request
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found

  /orders/stats:
    get:
      operationId: getOrderStats
      tags: [orders]
      summary: Get order statistics
      responses:
        "200":
          description: Order statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    MenuItem:
      type: object
      required: [id, name, nameAr, price, category]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        price:
          type: number
        category:
          type: string
        description:
          type: ["string", "null"]
        available:
          type: boolean

    Category:
      type: object
      required: [name, nameAr, count]
      properties:
        name:
          type: string
        nameAr:
          type: string
        count:
          type: integer

    OrderItem:
      type: object
      required: [menuItemId, quantity]
      properties:
        menuItemId:
          type: integer
        quantity:
          type: integer
        notes:
          type: ["string", "null"]

    OrderInput:
      type: object
      required: [items, customerName, phone]
      properties:
        customerName:
          type: string
        phone:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        notes:
          type: ["string", "null"]

    OrderItemDetail:
      type: object
      required: [menuItemId, menuItemName, menuItemNameAr, quantity, unitPrice, subtotal]
      properties:
        menuItemId:
          type: integer
        menuItemName:
          type: string
        menuItemNameAr:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        subtotal:
          type: number
        notes:
          type: ["string", "null"]

    Order:
      type: object
      required: [id, customerName, phone, status, total, createdAt, items]
      properties:
        id:
          type: integer
        customerName:
          type: string
        phone:
          type: string
        status:
          type: string
        total:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemDetail"

    OrderStats:
      type: object
      required: [totalOrders, totalRevenue, popularItems]
      properties:
        totalOrders:
          type: integer
        totalRevenue:
          type: number
        popularItems:
          type: array
          items:
            type: object
            required: [name, nameAr, count]
            properties:
              name:
                type: string
              nameAr:
                type: string
              count:
                type: integer