# lib/api-zod/src/generated/types/menuItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hamza Cafe API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MenuItem {
  id: number;
  name: string;
  nameAr: string;
  price: number;
  category: string;
  /** @nullable */
  description?: string | null;
  available?: boolean;
}