# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hamza Cafe API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItemDetail } from './orderItemDetail';

export interface Order {
  id: number;
  customerName: string;
  phone: string;
  status: string;
  total: number;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  items: OrderItemDetail[];
}