# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hamza Cafe API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';

export interface OrderInput {
  customerName: string;
  phone: string;
  items: OrderItem[];
  /** @nullable */
  notes?: string | null;
}