# lib/api-zod/src/generated/types/orderItemDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hamza Cafe API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItemDetail {
  menuItemId: number;
  menuItemName: string;
  menuItemNameAr: string;
  quantity: number;
  unitPrice: number;
  subtotal: number;
  /** @nullable */
  notes?: string | null;
}