# lib/api-zod/src/generated/types/orderStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hamza Cafe API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatsPopularItemsItem } from './orderStatsPopularItemsItem';

export interface OrderStats {
  totalOrders: number;
  totalRevenue: number;
  popularItems: OrderStatsPopularItemsItem[];
}